Final revision + changes
Code commenting, cleanup and preperation for submission
