+++ HomeHQ/Bill/BillView.swift
@@ -11,7 +11,11 @@ import SwiftUI
 // Not implemented
 struct BillView: View {
     var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+        ZStack {
+            Color("PrimaryBackground")
+                .edgesIgnoringSafeArea(.all)
+            Text("Not yet implemented")
+        }
     }
 }
