--- HomeHQ/Core/DashboardViewModel.swift
@@ -1,13 +1,0 @@
-//
-//  DashboardViewModel.swift
-//  HomeHQ
-//
-//  Created by Cooper Jacob on 23/8/2023.
-//
-
-import Foundation
-
-// View model for Dashboard view
-class DashboardViewModel: BaseViewModel {
-    
-}
