--- HomeHQ/Reminders/RemindersViewModel.swift
@@ -1,13 +1,0 @@
-//
-//  RemindersViewModel.swift
-//  HomeHQ
-//
-//  Created by Cooper Jacob on 27/8/2023.
-//
-
-import Foundation
-
-// View model for RemindersView
-final class RemindersViewModel: ObservableObject {
-    
-}
