--- HomeHQ/Settings/JoinHomeView.swift
@@ -1,20 +1,0 @@
-//
-//  JoinHomeView.swift
-//  HomeHQ
-//
-//  Created by Cooper Jacob on 11/9/2023.
-//
-
-import SwiftUI
-
-struct JoinHomeView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-struct JoinHomeView_Previews: PreviewProvider {
-    static var previews: some View {
-        JoinHomeView()
-    }
-}
