+++ HomeHQ/Tasks/HouseDutiesView.swift
@@ -10,8 +10,11 @@ import SwiftUI
 // Not implemented
 struct HouseDutiesView: View {
     var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
+        ZStack {
+            Color("PrimaryBackground")
+                .edgesIgnoringSafeArea(.all)
+            Text("Not yet implemented")
+        }    }
 }
 
 struct HouseDutiesView_Previews: PreviewProvider {
