+++ HomeHQ/Utils/ViewUtils.swift
@@ -10,6 +10,7 @@ import UIKit
 import SwiftUI
 
 // MARK: DO NOT MARK - WRITTEN PRIOR TO THIS SUBJECT
+
 final class ViewUtils {
     static let shared = ViewUtils()
     private init() { }
